Refactor RpcCall trait - create separate structs for each implementation
This had to be done in order to allow for different return values
from each specific RPC call implementation
