+++ enclave/src/rpc.rs
@@ -7,7 +7,11 @@ pub mod worker_api_direct;
 
 pub mod io_handler_extensions;
 pub mod return_value_encoding;
-pub mod rpc_api_calls;
-pub mod rpc_call;
 pub mod rpc_call_encoder;
 pub mod rpc_info;
+
+mod rpc_cancel_order;
+mod rpc_get_balance;
+mod rpc_place_order;
+mod rpc_withdraw;
+mod trusted_operation_verifier;
