+++ enclave/src/tests.rs
@@ -21,8 +21,8 @@ use crate::rpc;
 use crate::rsa3072;
 use crate::state;
 use crate::test_orderbook_storage;
-use crate::test_proxy;
 use crate::test_polkadex_gateway;
+use crate::test_proxy;
 use crate::top_pool;
 
 use crate::{Timeout, WorkerRequest, WorkerResponse};
@@ -65,7 +65,6 @@ use sp_core::ed25519 as spEd25519;
 
 use rpc::author::{Author, AuthorApi};
 use rpc::io_handler_extensions;
-use rpc::rpc_call;
 use rpc::rpc_call_encoder;
 use rpc::{api::SideChainApi, basic_pool::BasicPool};
 
@@ -138,8 +137,8 @@ pub extern "C" fn test_main_entrance() -> size_t {
         state::test_sgx_state_decode_encode_works,
         state::test_encrypt_decrypt_state_type_works,
         rpc_call_encoder::tests::test_encoding_none_params_returns_ok,
-        rpc_call::tests::test_method_name_should_not_be_empty,
-        rpc_call::tests::test_given_none_params_return_ok_result,
+        //rpc_call::tests::test_method_name_should_not_be_empty,
+        //rpc_call::tests::test_given_none_params_return_ok_result,
         io_handler_extensions::tests::test_given_io_handler_methods_then_retrieve_all_names_as_string,
         test_time_is_overdue,
         test_time_is_not_overdue,
